[CausalLM] Separate props from qwen_moe layer
- This patch moves the props for qwen_moe_layer from its header to
  common property header for causallm.
