+++ Applications/CausalLM/layers/causallm_common_properties.h
@@ -35,7 +35,36 @@
 #include <utility>
 
 namespace causallm {
+
 namespace props {
+
+/**
+ * @brief MoE activation type
+ */
+class MoEActivation final
+  : public nntrainer::EnumProperty<nntrainer::props::ActivationTypeInfo> {
+public:
+  using prop_tag = nntrainer::enum_class_prop_tag;
+  static constexpr const char *key = "moe_activation";
+};
+/**
+ * @brief NumExperts,  Number of experts property
+ */
+class NumExperts : public nntrainer::PositiveIntegerProperty {
+public:
+  static constexpr const char *key = "num_experts"; /**< unique key to access */
+  using prop_tag = nntrainer::uint_prop_tag;        /**< property type */
+};
+
+/**
+ * @brief NumExpertsPerToken,  Number of experts per token property
+ */
+class NumExpertsPerToken : public nntrainer::PositiveIntegerProperty {
+public:
+  static constexpr const char *key =
+    "num_experts_per_token";                 /**< unique key to access */
+  using prop_tag = nntrainer::uint_prop_tag; /**< property type */
+};
 
 /**
  * @brief unit property, unit is used to measure how many weights are there
